Restructured fb update components
